issue #717 separate cam demo routing
